+++ packages/remote-config/src/abt/experiment.ts
@@ -28,7 +28,8 @@ export class Experiment {
   async updateActiveExperiments(
     latestExperiments: FirebaseExperimentDescription[]
   ): Promise<void> {
-    const currentActiveExperiments = await this.storage.getActiveExperiments() || new Set<string>();
+    const currentActiveExperiments =
+      (await this.storage.getActiveExperiments()) || new Set<string>();
     const experimentInfoMap = this.createExperimentInfoMap(latestExperiments);
     this.addActiveExperiments(currentActiveExperiments, experimentInfoMap);
     this.removeInactiveExperiments(currentActiveExperiments, experimentInfoMap);
@@ -47,7 +48,8 @@ export class Experiment {
 
   private addActiveExperiments(
     currentActiveExperiments: Set<string>,
-    experimentInfoMap: Map<string, FirebaseExperimentDescription>): void {
+    experimentInfoMap: Map<string, FirebaseExperimentDescription>
+  ): void {
     for (const [experimentId, experimentInfo] of experimentInfoMap.entries()) {
       if (!currentActiveExperiments.has(experimentId)) {
         this.addExperimentToAnalytics(experimentId, experimentInfo.variantId);
@@ -66,13 +68,14 @@ export class Experiment {
     }
   }
 
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  private addExperimentToAnalytics(experimentId: string, variantId: string): void {
+  private addExperimentToAnalytics(
+    _experimentId: string,
+    _variantId: string
+  ): void {
     // TODO
   }
 
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  private removeExperimentFromAnalytics(experimentId: string): void {
+  private removeExperimentFromAnalytics(_experimentId: string): void {
     // TODO
   }
 }