+++ packages/remote-config/src/client/rest_client.ts
@@ -89,6 +89,8 @@ export class RestClient implements RemoteConfigFetchClient {
       // Deviates from pure decorator by not passing max-age header since we don't currently have
       // service behavior using that header.
       'If-None-Match': request.eTag || '*'
+      // TODO: Add this header once CORS error is fixed internally.
+      //'X-Firebase-RC-Fetch-Type': `${fetchType}/${fetchAttempt}`
     };
 
     const requestBody: FetchRequestBody = {
@@ -141,6 +143,7 @@ export class RestClient implements RemoteConfigFetchClient {
 
     let config: FirebaseRemoteConfigObject | undefined;
     let state: string | undefined;
+    let templateVersion: number | undefined;
     let experiments: FirebaseExperimentDescription[] | undefined;
 
     // JSON parsing throws SyntaxError if the response body isn't a JSON string.
@@ -156,6 +159,7 @@ export class RestClient implements RemoteConfigFetchClient {
       }
       config = responseBody['entries'];
       state = responseBody['state'];
+      templateVersion = responseBody['templateVersion'];
       experiments = responseBody['experimentDescriptions'];
     }
 
@@ -180,6 +184,6 @@ export class RestClient implements RemoteConfigFetchClient {
       });
     }
 
-    return { status, eTag: responseEtag, config, experiments };
+    return { status, eTag: responseEtag, config, templateVersion, experiments };
   }
 }