+++ packages/remote-config/src/errors.ts
@@ -33,7 +33,11 @@ export const enum ErrorCode {
   FETCH_PARSE = 'fetch-client-parse',
   FETCH_STATUS = 'fetch-status',
   INDEXED_DB_UNAVAILABLE = 'indexed-db-unavailable',
-  CUSTOM_SIGNAL_MAX_ALLOWED_SIGNALS = 'custom-signal-max-allowed-signals'
+  CUSTOM_SIGNAL_MAX_ALLOWED_SIGNALS = 'custom-signal-max-allowed-signals',
+  CONFIG_UPDATE_STREAM_ERROR = 'stream-error',
+  CONFIG_UPDATE_UNAVAILABLE = 'realtime-unavailable',
+  CONFIG_UPDATE_MESSAGE_INVALID = 'update-message-invalid',
+  CONFIG_UPDATE_NOT_FETCHED = 'update-not-fetched'
 }
 
 const ERROR_DESCRIPTION_MAP: { readonly [key in ErrorCode]: string } = {
@@ -72,7 +76,15 @@ const ERROR_DESCRIPTION_MAP: { readonly [key in ErrorCode]: string } = {
   [ErrorCode.INDEXED_DB_UNAVAILABLE]:
     'Indexed DB is not supported by current browser',
   [ErrorCode.CUSTOM_SIGNAL_MAX_ALLOWED_SIGNALS]:
-    'Setting more than {$maxSignals} custom signals is not supported.'
+    'Setting more than {$maxSignals} custom signals is not supported.',
+  [ErrorCode.CONFIG_UPDATE_STREAM_ERROR]:
+    'The stream was not able to connect to the backend: {$originalErrorMessage}.',
+  [ErrorCode.CONFIG_UPDATE_UNAVAILABLE]:
+    'The Realtime service is unavailable: {$originalErrorMessage}',
+  [ErrorCode.CONFIG_UPDATE_MESSAGE_INVALID]:
+    'The stream invalidation message was unparsable: {$originalErrorMessage}',
+  [ErrorCode.CONFIG_UPDATE_NOT_FETCHED]:
+    'Unable to fetch the latest config: {$originalErrorMessage}'
 };
 
 // Note this is effectively a type system binding a code to params. This approach overlaps with the
@@ -92,6 +104,10 @@ interface ErrorParams {
   [ErrorCode.FETCH_PARSE]: { originalErrorMessage: string };
   [ErrorCode.FETCH_STATUS]: { httpStatus: number };
   [ErrorCode.CUSTOM_SIGNAL_MAX_ALLOWED_SIGNALS]: { maxSignals: number };
+  [ErrorCode.CONFIG_UPDATE_STREAM_ERROR]: { originalErrorMessage: string };
+  [ErrorCode.CONFIG_UPDATE_UNAVAILABLE]: { originalErrorMessage: string };
+  [ErrorCode.CONFIG_UPDATE_MESSAGE_INVALID]: { originalErrorMessage: string };
+  [ErrorCode.CONFIG_UPDATE_NOT_FETCHED]: { originalErrorMessage: string };
 }
 
 export const ERROR_FACTORY = new ErrorFactory<ErrorCode, ErrorParams>(