+++ packages/remote-config/src/storage/storage.ts
@@ -56,6 +56,13 @@ export interface ThrottleMetadata {
   throttleEndTimeMillis: number;
 }
 
+export interface RealtimeBackoffMetadata {
+  // The number of consecutive connection streams that have failed.
+  numFailedStreams: number;
+  // The Date until which the client should wait before attempting any new real-time connections.
+  backoffEndTimeMillis: Date;
+}
+
 /**
  * Provides type-safety for the "key" field used by {@link APP_NAMESPACE_STORE}.
  *
@@ -70,7 +77,9 @@ type ProjectNamespaceKeyFieldValue =
   | 'last_successful_fetch_response'
   | 'settings'
   | 'throttle_metadata'
-  | 'custom_signals';
+  | 'custom_signals'
+  | 'realtime_backoff_metadata'
+  | 'last_known_template_version';
 
 // Visible for testing.
 export function openDatabase(): Promise<IDBDatabase> {
@@ -161,7 +170,7 @@ export abstract class Storage {
     return this.get<Set<string>>('active_experiments');
   }
 
-  setActiveExperiments(experiments:Set<string>): Promise<void> {
+  setActiveExperiments(experiments: Set<string>): Promise<void> {
     return this.set<Set<string>>('active_experiments', experiments);
   }
 
@@ -187,6 +196,27 @@ export abstract class Storage {
   abstract get<T>(key: ProjectNamespaceKeyFieldValue): Promise<T | undefined>;
   abstract set<T>(key: ProjectNamespaceKeyFieldValue, value: T): Promise<void>;
   abstract delete(key: ProjectNamespaceKeyFieldValue): Promise<void>;
+
+  getRealtimeBackoffMetadata(): Promise<RealtimeBackoffMetadata | undefined> {
+    return this.get<RealtimeBackoffMetadata>('realtime_backoff_metadata');
+  }
+
+  setRealtimeBackoffMetadata(
+    realtimeMetadata: RealtimeBackoffMetadata
+  ): Promise<void> {
+    return this.set<RealtimeBackoffMetadata>(
+      'realtime_backoff_metadata',
+      realtimeMetadata
+    );
+  }
+
+  getActiveConfigTemplateVersion(): Promise<number | undefined> {
+    return this.get<number>('last_known_template_version');
+  }
+
+  setActiveConfigTemplateVersion(version: number): Promise<void> {
+    return this.set<number>('last_known_template_version', version);
+  }
 }
 
 export class IndexedDbStorage extends Storage {