+++ packages/remote-config/test/client/rest_client.test.ts
@@ -26,6 +26,7 @@ import {
   FetchRequest,
   RemoteConfigAbortSignal
 } from '../../src/client/remote_config_fetch_client';
+import { Storage } from '../../src/storage/storage';
 
 const DEFAULT_REQUEST: FetchRequest = {
   cacheMaxAgeMillis: 1,
@@ -34,6 +35,7 @@ const DEFAULT_REQUEST: FetchRequest = {
 
 describe('RestClient', () => {
   const firebaseInstallations = {} as FirebaseInstallations;
+  const storage = {} as Storage;
   let client: RestClient;
 
   beforeEach(() => {
@@ -51,6 +53,7 @@ describe('RestClient', () => {
     firebaseInstallations.getToken = sinon
       .stub()
       .returns(Promise.resolve('fis-token'));
+    storage.setActiveConfigTemplateVersion = sinon.stub();
   });
 
   describe('fetch', () => {
@@ -75,13 +78,16 @@ describe('RestClient', () => {
         eTag: 'etag',
         state: 'UPDATE',
         entries: { color: 'sparkling' },
-        experimentDescriptions: [{
-          experimentId: "_exp_1",
-          variantId : "1",
-          experimentStartTime : "2025-04-06T14:13:57.597Z",
-          triggerTimeoutMillis : "15552000000",
-          timeToLiveMillis : "15552000000"
-        }]
+        templateVersion: 1,
+        experimentDescriptions: [
+          {
+            experimentId: '_exp_1',
+            variantId: '1',
+            experimentStartTime: '2025-04-06T14:13:57.597Z',
+            triggerTimeoutMillis: '15552000000',
+            timeToLiveMillis: '15552000000'
+          }
+        ]
       };
 
       fetchStub.returns(
@@ -93,6 +99,7 @@ describe('RestClient', () => {
             Promise.resolve({
               entries: expectedResponse.entries,
               state: expectedResponse.state,
+              templateVersion: expectedResponse.templateVersion,
               experimentDescriptions: expectedResponse.experimentDescriptions
             })
         } as Response)
@@ -104,6 +111,7 @@ describe('RestClient', () => {
         status: expectedResponse.status,
         eTag: expectedResponse.eTag,
         config: expectedResponse.entries,
+        templateVersion: expectedResponse.templateVersion,
         experiments: expectedResponse.experimentDescriptions
       });
     });
@@ -194,6 +202,7 @@ describe('RestClient', () => {
         status: 304,
         eTag: 'response-etag',
         config: undefined,
+        templateVersion: undefined,
         experiments: undefined
       });
     });
@@ -233,6 +242,7 @@ describe('RestClient', () => {
         status: 304,
         eTag: 'etag',
         config: undefined,
+        templateVersion: undefined,
         experiments: undefined
       });
     });
@@ -251,6 +261,7 @@ describe('RestClient', () => {
           status: 200,
           eTag: 'etag',
           config: {},
+          templateVersion: undefined,
           experiments: []
         });
       }