+++ packages/app/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @firebase/app
 
+## 0.14.3
+
+### Patch Changes
+
+- Update SDK_VERSION.
+
+## 0.14.2
+
+### Patch Changes
+
+- Update SDK_VERSION.
+
+## 0.14.1
+
+### Patch Changes
+
+- Update SDK_VERSION.
+
 ## 0.14.0
 
 ### Minor Changes